requirements.txt
celery
Django==1.11
django-crispy-forms
django-simple-captcha
Markdown
redis
python-bitcoinrpc
